requirements.txt
# DataFlip MX - Dependencias del Proyecto
# Actualizado: Diciembre 2024

# === CORE ===
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# === WEB SCRAPING & APIs ===
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
lxml>=4.9.0

# === GOOGLE TRENDS ===
pytrends>=4.9.0

# === REDDIT API ===
praw>=7.7.0

# === DATA ANALYSIS ===
scikit-learn>=1.3.0
scipy>=1.11.0

# === VISUALIZACIÓN ===
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# === JUPYTER ===
jupyter>=1.0.0
ipykernel>=6.25.0

# === UTILIDADES ===
tqdm>=4.66.0  # Progress bars
openpyxl>=3.1.0  # Para exportar Excel
